requirements.txt
flask
python-dotenv
azure-ai-formrecognizer
requests
